requirements.txt
streamlit==1.24.1
youtube-transcript-api==0.6.0
google-generativeai==0.8.5
python-dotenv==1.0.0
langcodes==3.3.0
mypy==1.3.0
